+++ src/settings.ts
@@ -1,4 +1,5 @@
 export const MODULE_NAME = "shared-token-visibility";
 export enum SettingName {
     RevealNpc = "revealNpcs",
+    PlayerVision = "playerVision",
 }