+++ src/types/Token.d.ts
@@ -1,0 +1,5 @@
+declare class Token extends PlaceableObject {
+    actor: Actor | undefined;
+    hasSight: boolean;
+    _isVisionSource(): boolean;
+}