+++ CHANGELOG.md
@@ -10,7 +10,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 -   Token bars and effects are now rendered in addition to the icon.
--   The module now has settings. First new setting: You can now opt out of sharing NPC token visibility (making this module only share player tokens).
+-   The module now has settings:
+    1. You can now opt out of sharing NPC token visibility (making this module only share player tokens).
+    2. You can now enable shared _vision_ between all players. When this is active as a GM, clicking a player will show you combined player vision.
 
 ### Fixed